+++ src/main/java/net/sirplop/aetherworks/blockentity/render/RenderForge.java
@@ -17,14 +17,15 @@ public class RenderForge implements BlockEntityRenderer<ForgeCoreBlockEntity> {
     public RenderForge(BlockEntityRendererProvider.Context context)  { }
 
     public static final ResourceLocation LOCATION_FORGE_OVERLAY = new ResourceLocation(Aetherworks.MODID, "textures/block/forge/overlay.png");
-    public static final float[] COLOR = {1, 1, 1, 1};
+    //public static final float[] COLOR = {1, 1, 1, 1};
     public static final int[] LIGHTMAP = {240, 240};
 
-    private static final int frameTime = 20;
+    private static final int frameTime = 3;
     private static final int frameCount = 20;
     private static final float frameUVShift = 1f / frameCount;
     private int frame = 0;
     private float time = 0;
+    private float animTime = 0;
     private boolean flip = false;
 
     @Override
@@ -34,21 +35,16 @@ public class RenderForge implements BlockEntityRenderer<ForgeCoreBlockEntity> {
             return;
         }
         time += pPartialTick;
-        if (time >= frameTime)
-        {
-            time = 0;
-            if (flip)
-            {
-                if (--frame <= -1)
-                {
+        animTime += pPartialTick;
+        if (animTime >= frameTime) {
+            animTime = 0;
+            if (flip) {
+                if (--frame <= -1) {
                     flip = false;
                     frame += 2;
                 }
-            }
-            else
-            {
-                if (++frame >= frameCount)
-                {
+            } else {
+                if (++frame >= frameCount) {
                     flip = true;
                     frame -= 2;
                 }
@@ -61,26 +57,29 @@ public class RenderForge implements BlockEntityRenderer<ForgeCoreBlockEntity> {
         float uv0 = frameUVShift * frame;
         float uv1 = frameUVShift * (frame + 1);
 
+        float sin = 0.5F + (float) Math.abs(Math.sin(Math.toRadians(time * 3))) * 0.5F;
+        float[] color = {1, sin, sin, 1};
+
         //render all 12 faces of the forge
         poseStack.translate(-1, -1, -1); //north-west corner
-        FaceRendererUtil.renderFace(Direction.NORTH, overlayBuffer, poseStack, COLOR, LIGHTMAP, 0, 1, uv0, uv1);
-        FaceRendererUtil.renderFace(Direction.WEST, overlayBuffer, poseStack, COLOR, LIGHTMAP, 0, 1, uv0, uv1);
+        FaceRendererUtil.renderFace(Direction.NORTH, overlayBuffer, poseStack, color, LIGHTMAP, 0, 1, uv0, uv1);
+        FaceRendererUtil.renderFace(Direction.WEST, overlayBuffer, poseStack, color, LIGHTMAP, 0, 1, uv0, uv1);
         poseStack.translate(1, 0, 0); //north side
-        FaceRendererUtil.renderFace(Direction.NORTH, overlayBuffer, poseStack, COLOR, LIGHTMAP, 0, 1, uv0, uv1);
+        FaceRendererUtil.renderFace(Direction.NORTH, overlayBuffer, poseStack, color, LIGHTMAP, 0, 1, uv0, uv1);
         poseStack.translate(1, 0, 0); //north-east corner
-        FaceRendererUtil.renderFace(Direction.NORTH, overlayBuffer, poseStack, COLOR, LIGHTMAP, 0, 1, uv0, uv1);
-        FaceRendererUtil.renderFace(Direction.EAST, overlayBuffer, poseStack, COLOR, LIGHTMAP, 0, 1, uv0, uv1);
+        FaceRendererUtil.renderFace(Direction.NORTH, overlayBuffer, poseStack, color, LIGHTMAP, 0, 1, uv0, uv1);
+        FaceRendererUtil.renderFace(Direction.EAST, overlayBuffer, poseStack, color, LIGHTMAP, 0, 1, uv0, uv1);
         poseStack.translate(0, 0, 1); //east side
-        FaceRendererUtil.renderFace(Direction.EAST, overlayBuffer, poseStack, COLOR, LIGHTMAP, 0, 1, uv0, uv1);
+        FaceRendererUtil.renderFace(Direction.EAST, overlayBuffer, poseStack, color, LIGHTMAP, 0, 1, uv0, uv1);
         poseStack.translate(0, 0, 1); //south-east corner
-        FaceRendererUtil.renderFace(Direction.SOUTH, overlayBuffer, poseStack, COLOR, LIGHTMAP, 0, 1, uv0, uv1);
-        FaceRendererUtil.renderFace(Direction.EAST, overlayBuffer, poseStack, COLOR, LIGHTMAP, 0, 1, uv0, uv1);
+        FaceRendererUtil.renderFace(Direction.SOUTH, overlayBuffer, poseStack, color, LIGHTMAP, 0, 1, uv0, uv1);
+        FaceRendererUtil.renderFace(Direction.EAST, overlayBuffer, poseStack, color, LIGHTMAP, 0, 1, uv0, uv1);
         poseStack.translate(-1, 0, 0); //south side
-        FaceRendererUtil.renderFace(Direction.SOUTH, overlayBuffer, poseStack, COLOR, LIGHTMAP, 0, 1, uv0, uv1);
+        FaceRendererUtil.renderFace(Direction.SOUTH, overlayBuffer, poseStack, color, LIGHTMAP, 0, 1, uv0, uv1);
         poseStack.translate(-1, 0, 0); //south-west corner
-        FaceRendererUtil.renderFace(Direction.SOUTH, overlayBuffer, poseStack, COLOR, LIGHTMAP, 0, 1, uv0, uv1);
-        FaceRendererUtil.renderFace(Direction.WEST, overlayBuffer, poseStack, COLOR, LIGHTMAP, 0, 1, uv0, uv1);
+        FaceRendererUtil.renderFace(Direction.SOUTH, overlayBuffer, poseStack, color, LIGHTMAP, 0, 1, uv0, uv1);
+        FaceRendererUtil.renderFace(Direction.WEST, overlayBuffer, poseStack, color, LIGHTMAP, 0, 1, uv0, uv1);
         poseStack.translate(0, 0, -1); //west side
-        FaceRendererUtil.renderFace(Direction.WEST, overlayBuffer, poseStack, COLOR, LIGHTMAP, 0, 1, uv0, uv1);
+        FaceRendererUtil.renderFace(Direction.WEST, overlayBuffer, poseStack, color, LIGHTMAP, 0, 1, uv0, uv1);
     }
 }